requirements.txt
numpy==2.2.2
pandas==2.2.3
scipy==1.15.1
pybedtools==0.10.0
notebook==7.3.2
pydantic==2.10.6
numpydantic==1.6.7
pyranges1==1.0.3
s3fs==2025.2.0
fsspec~=2025.2.0
temporalio==1.10.0
dotenv~=0.9.9
python-dotenv~=1.0.1
pymongo~=4.8.0
pyarrow~=19.0.0
